--- go.mod
@@ -1,9 +1,0 @@
-module supply
-
-go 1.13
-
-require (
-	github.com/go-kit/kit v0.9.0
-	github.com/go-logfmt/logfmt v0.5.0 // indirect
-	github.com/gorilla/mux v1.7.3
-)
