+++ pilot-management/domain/entity/pilot.go
@@ -1,0 +1,10 @@
+package entity
+
+type Pilot struct {
+	Id         string
+	UserId     string
+	CodeName   string
+	SupplierId string
+	MarketId   string
+	ServiceId  string
+}
