+++ pilot-management/go.mod
@@ -1,0 +1,12 @@
+module gitlab.com/afmelsaidy/pilot-management
+
+replace gitlab.com/afmelsaidy/pilot-management => ./
+
+go 1.13
+
+require (
+	github.com/go-kit/kit v0.9.0
+	github.com/go-logfmt/logfmt v0.5.0 // indirect
+	github.com/go-stack/stack v1.8.0 // indirect
+	github.com/gorilla/mux v1.7.3
+)
