+++ pilot-management/repository/impl/mock/pilot.go
@@ -1,0 +1,11 @@
+package mock
+
+import "../../../domain/entity"
+
+type PilotRepoMock struct {
+	store map[string]entity.Pilot
+}
+
+func (r *PilotRepoMock) ListPilots() ([]entity.Pilot, error) {
+	return []entity.Pilot{}, nil
+}
