--- pilot/response.go
@@ -1,6 +1,0 @@
-package pilot
-
-type Response struct {
-	data interface {}
-	errors []error
-}
