--- pilot/view.go
@@ -1,21 +1,0 @@
-package pilot
-
-type PilotView struct {
-	Id         string `json:"id"`
-	UserId     string `json:"userId"`
-	CodeName   string `json:"codeName"`
-	SupplierId string `json:"supplierId"`
-	MarketId   string `json:"marketId"`
-	ServiceId  string `json:"serviceId"`
-}
-
-func toPilotView(pilot Pilot) PilotView {
-	return PilotView{
-		Id:         pilot.Id,
-		UserId:     pilot.UserId,
-		CodeName:   pilot.CodeName,
-		SupplierId: pilot.ServiceId,
-		MarketId:   pilot.MarketId,
-		ServiceId:  pilot.ServiceId,
-	}
-}
